feat: SQLite persistence, Docker deployment, bug fixes
- New db.py: SQLite persistence for processed emails, pending expenses,
  pending email expenses, debit queue, conversation history
- Removes in-memory ConversationState, all state survives restarts
- Auto-migration from processed_emails.json on first run
- agent.py: answer_query takes history list, returns updated history
- email_watcher.py: all persistence via DB, removes JSON file handling
- main.py: DB lifecycle (init/close), handle_cancel auth guard,
  history append instead of clear+rewrite
- Dockerfile + docker-compose.yaml for deployment
- .dockerignore + updated .gitignore
- Fixes: handle_cancel missing auth guard, notion.py type hint,
  dead variable in prune_processed, owner_name regression

+++ notion.py
@@ -56,17 +56,17 @@ class NotionClient:
         self,
         subcategories: dict[str, str],
         accounts: dict[str, str],
-    ) -> dict[float, dict]:
+    ) -> dict[int, dict]:
         """
         Load Active entries from the Recurring Payment database.
-        Returns dict keyed by amount (IDR float) → {name, page_url, subcategory, account}.
+        Returns dict keyed by amount (IDR int) → {name, page_url, subcategory, account}.
         """
         # Build reverse maps: Notion page ID → human-readable name
         sub_id_to_name = {_url_to_id(url): name for name, url in subcategories.items()}
         acc_id_to_name = {_url_to_id(url): name for name, url in accounts.items()}
 
         pages = await self._query_db(self._config.recurring_ds)
-        result: dict[float, dict] = {}
+        result: dict[int, dict] = {}
 
         for p in pages:
             # Only Active entries
